-added the option to use virtual subfolders (or tiers) to the source menu. works in sourceflow and regular source menu. if you wish to do this you need to edit your source_menu.ini and do the following.
add src_folder=x to each button. x is the number of the virtual folder this button belongs in. 0 is obviously the base folder.

add source=folder when you want a button to open a new virtual folder and folder=x to specify which new folder to open. in folders 1 and above you will need to make a back or up button to go back to the previous folder.